+++ src/main/webapp/WEB-INF/views/bbs/review_update.jsp
@@ -150,17 +150,17 @@ fieldset {
 													<td bgcolor="#1b5ac2" class="w_font">물품</td>
 													<c:choose>
 													<c:when test="${empty reviewvo.PROD_NAME}">
-														<td id="text" style="color:gray; font-size:15px;">[선택물품없음]</td>
+													<td><input type="text" name="PROD_NAME" size="20" autocomplete='off' value="[선택물품없음]" disabled/></td>
 													</c:when>
 													<c:otherwise>
-														<td id="text">${reviewvo.PROD_NAME}</td>
+													<td><input type="text" name="PROD_NAME" size="20" autocomplete='off' value="${reviewvo.PROD_NAME}" disabled/></td>
 													</c:otherwise>
 												</c:choose>
 												</tr>
 										<tr align="center">
 								<td bgcolor="#1b5ac2" class="w_font">작성자</td>
 								<!--이건 로그인한 사람이 자동으로 뜨게하기.  -->
-								<td id="text">${reviewvo.RE_WRITER}</td>
+								<td><input type="text" name="RE_WRITER" size="20" autocomplete='off' value="${reviewvo.RE_WRITER}" disabled/></td>
 							</tr>
 							<tr align="center">
 								<td bgcolor="#1b5ac2" class="w_font">제목</td>
@@ -172,7 +172,7 @@ fieldset {
 						
 								  <c:choose>
 									  	<c:when test="${empty reviewvo.RE_FILE}">
-									  		<td><input type="file" name="file"><br><b>이전 파일 없음</b></td>
+									  		<td><input type="file" name="file"><b>이전 파일 없음</b></td>
 									  		<input type="hidden" name="old_f_name" value="">
 									  	</c:when>
 									  	<c:otherwise>
